+++ src/screens/ScreenOffense.tsx
@@ -3,6 +3,10 @@ import { ReactNode, useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
 import { useSessionStorage } from "usehooks-ts";
+import {
+  CheckboxGroup,
+  CheckboxGroupOption,
+} from "../components/CheckboxGroup";
 import { CopyButton } from "../components/CopyButton";
 import { DexCoverage } from "../components/DexCoverage";
 import { FancyLink } from "../components/FancyLink";
@@ -13,9 +17,13 @@ import { MultiTypeSelector } from "../components/MultiTypeSelector";
 import { useAppContext } from "../hooks/useAppContext";
 import { useGeneration } from "../hooks/useGeneration";
 import { useSearch } from "../hooks/useSearch";
+import { compare } from "../misc/compare";
 import {
+  AbilityName,
+  SpecialMove,
   Type,
   removeInvalidOffenseTypesForGeneration,
+  splitTokens,
   typesFromString,
 } from "../misc/data-types";
 import styles from "./ScreenOffense.module.css";
@@ -31,14 +39,15 @@ export function ScreenOffense(): ReactNode {
     []
   );
 
-  const permalink = new URL(window.location.href);
-  if (offenseTypes.length > 0) {
-    permalink.searchParams.set("types", offenseTypes.join(" "));
-  }
-
   useEffect(() => {
     if (search.has("types")) {
       setOffenseTypes(typesFromString(search.get("types") || ""));
+    }
+    if (search.has("moves")) {
+      setSpecialMoves(splitTokens(search.get("moves") || "") as any);
+    }
+    if (search.has("abilities")) {
+      setAbilities(splitTokens(search.get("abilities") || "") as any);
     }
     navigate({ search: "" }, { replace: true });
   }, [search]);
@@ -52,6 +61,49 @@ export function ScreenOffense(): ReactNode {
   const listLength = coverageTypes?.length ?? 0;
   const listLengthFormatted = listLength.toLocaleString(i18n.languages);
 
+  const specialMovesOptions: CheckboxGroupOption<SpecialMove>[] = [
+    {
+      id: "thousand_arrows",
+      name: t(`offense.specialMoves.names.thousand_arrows`),
+    },
+    { id: "freeze-dry", name: t(`offense.specialMoves.names.freeze-dry`) },
+  ] as const;
+  specialMovesOptions.sort((a, b) => {
+    return compare(a.name, b.name);
+  });
+
+  const [specialMoves, setSpecialMoves] = useSessionStorage<SpecialMove[]>(
+    "offense.specialMoves",
+    []
+  );
+
+  const abilitiesOptions: CheckboxGroupOption<AbilityName>[] = [
+    {
+      id: "tinted_lens",
+      name: t(`defense.abilityNames.tinted_lens`),
+    },
+    { id: "scrappy", name: t(`defense.abilityNames.scrappy`) },
+  ] as const;
+  abilitiesOptions.sort((a, b) => {
+    return compare(a.name, b.name);
+  });
+
+  const [abilities, setAbilities] = useSessionStorage<AbilityName[]>(
+    "offense.abilities",
+    []
+  );
+
+  const permalink = new URL(window.location.href);
+  if (offenseTypes.length > 0) {
+    permalink.searchParams.set("types", offenseTypes.join(" "));
+  }
+  if (abilities.length > 0) {
+    permalink.searchParams.set("abilities", abilities.join(" "));
+  }
+  if (specialMoves.length > 0) {
+    permalink.searchParams.set("moves", specialMoves.join(" "));
+  }
+
   return (
     <main className={classNames(styles.root, "content-wide center")}>
       <Flex direction="column" gap="xlarge">
@@ -63,6 +115,26 @@ export function ScreenOffense(): ReactNode {
             generation={generation}
             value={offenseTypes}
             onChange={setOffenseTypes}
+          />
+        </Flex>
+        <Flex direction="column" gap="small">
+          <FancyText tag="h2" fontSize="large" fontWeight="medium">
+            {t("offense.specialMoves.choose")}
+          </FancyText>
+          <CheckboxGroup
+            options={specialMovesOptions}
+            value={specialMoves}
+            onChange={setSpecialMoves}
+          />
+        </Flex>
+        <Flex direction="column" gap="small">
+          <FancyText tag="h2" fontSize="large" fontWeight="medium">
+            {t("offense.abilities.choose")}
+          </FancyText>
+          <CheckboxGroup
+            options={abilitiesOptions}
+            value={abilities}
+            onChange={setAbilities}
           />
         </Flex>
         {generation === "default" && (
@@ -82,6 +154,8 @@ export function ScreenOffense(): ReactNode {
                 coverageTypes={coverageTypes ?? fallbackCoverageTypes}
                 types={offenseTypes}
                 isLoading={isLoading}
+                offenseAbilities={abilities}
+                specialMoves={specialMoves}
               />
               <Flex>
                 <CopyButton text={permalink.href}>
@@ -99,6 +173,8 @@ export function ScreenOffense(): ReactNode {
           types={offenseTypes}
           ability="none"
           teraType={Type.none}
+          specialMoves={specialMoves}
+          offenseAbilities={abilities}
         />
       </Flex>
     </main>